requirements.txt
aiogram==3.15.0
openai==1.59.5
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0